[ODLG] Migration DB.

+++ db/migrate/20120515002118_change_data_type_for_folios_activo.rb
@@ -1,7 +1,7 @@
 class ChangeDataTypeForFoliosActivo < ActiveRecord::Migration
   def up
     change_table :folios do |t|
-      t.change :activo, :float
+      t.change :activo, :integer
     end
   end
